feat: add data merge script and ensemble training support
- Add merge_data.py: combine Excel sheets + join DMP behavior data
- Add train_model.py: unified training router with train/test split modes
- Add train_test_drive_ensemble.py: 7/14/21 day test drive prediction ensemble
- Add validate_ensemble.py: ensemble model validation
- Add src/inference/: HAB derivation logic for three-model mode
- Update docs: QUICKSTART.md, CLAUDE.md with new commands
- Add polars dependency for efficient large file processing

+++ src/evaluation/ohab_metrics.py
@@ -144,7 +144,8 @@ def apply_hab_decision_policy(
         elif "A" in proba_df.columns:
             predictions.append("A")
         else:
-            predictions.append(row.astype(float).idxmax())
+            # 兜底逻辑：若无 A 则取概率最大的类
+            predictions.append(row.idxmax())
 
     return pd.Series(predictions, index=proba_df.index)
